Made the code be contained in a separate "approach" class
