+++ src/main/java/graphql/execution/instrumentation/dataloader/DataLoaderDispatcherInstrumentationState.java
@@ -1,0 +1,20 @@
+package graphql.execution.instrumentation.dataloader;
+
+import graphql.execution.instrumentation.InstrumentationState;
+
+/**
+ * A base class that keeps track of whether aggressive batching can be used
+ */
+public class DataLoaderDispatcherInstrumentationState implements InstrumentationState {
+    private boolean aggressivelyBatching = true;
+
+    boolean isAggressivelyBatching() {
+        return aggressivelyBatching;
+    }
+
+    void setAggressivelyBatching(boolean aggressivelyBatching) {
+        this.aggressivelyBatching = aggressivelyBatching;
+    }
+
+
+}
